requirements.txt
boto3
markovify
python-twitter
wikipedia
